gastos_cartao_corporativo/requirements.txt
numpy==1.23.3
pandas==1.5.0
Pillow==9.4.0
plotly==5.11.0
streamlit==1.17.0
